assignment 5: minor tweaks

+++ Source/ActionRoguelike/Public/AI/SAICharacter.h
@@ -40,7 +40,7 @@ protected:
 	TSubclassOf<UUserWidget> HealthBarWidgetClass;
 
 	UPROPERTY(EditDefaultsOnly, BlueprintReadWrite, Category = "Properties")
-	float CreditsValue;
+	int32 CreditsValue;
 	
 	USWorldUserWidget* ActiveHealthBar;
